move settings.js to config directory
